requirements.txt
# requirements.txt

numpy>=1.21,<2.0
pandas>=1.3,<2.0
PyYAML>=6.0,<7.0
yfinance>=0.2.18,<0.3.0
statsmodels>=0.13,<1.0
tensorflow>=2.10,<3.0
pytest>=7.0,<8.0
